Improves the UI.
[logger]
Improves teh logger to show the time stamp of the logs and to  use a
_logcontroller of type BehaviorSubject such that the logs can be updated
dynamically to any client that subscribes to it's services.

[peer_discovery_service] [server_service] [transfer_service]
Updates the way the use the logger.

[loge_panel]
Improves the UI of the log panel and also makes it collasable.

[file_explorer]
Improves teh UI of the file explorer.